fix ItemProperty
