+++ src/components/ItemProperty/ItemProperty.css
@@ -1,0 +1,25 @@
+.item-property-body-div {
+  display: flex;
+  flex-direction: column;
+  flex-wrap: nowrap;
+  align-items: start;
+  justify-content: flex-end;
+}
+
+.item-property-textarea-div {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: nowrap;
+  align-items: normal;
+  justify-content: space-around;
+  width: 100%;
+}
+
+.item-property-input-div {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: nowrap;
+  align-items: normal;
+  justify-content: space-around;
+  width: 100%;
+}
