--- src/components/ProductCategoryForm/index.tsx
@@ -1,3 +1,0 @@
-import ProductCategoryForm from "./ProductCategoryForm";
-
-export default ProductCategoryForm;
