--- src/components/ProductFeatureForm/ProductFeatureForm.css
@@ -1,22 +1,0 @@
-.product-feature-body-div {
-    display: flex;
-    flex-direction: column;
-    flex-wrap: nowrap;
-    align-items: start;
-    justify-content: flex-end;
-    /*height: 100%;*/
-    margin-left: 20px;
-}
-
-.product-feature-textarea-div {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: nowrap;
-    align-items: normal;
-    justify-content: space-around;
-    width: 100%;
-}
-
-.product-feature-button {
-    margin: 0.7rem 0 0 1rem;
-}
