--- src/components/ProductFeatureForm/index.tsx
@@ -1,3 +1,0 @@
-import ProductFeatureForm from "./ProductFeatureForm";
-
-export default ProductFeatureForm;
