--- src/components/ProductNameForm/index.tsx
@@ -1,3 +1,0 @@
-import ProductNameForm from "./ProductNameForm";
-
-export default ProductNameForm;
