--- src/components/ProductUrlForm/index.tsx
@@ -1,3 +1,0 @@
-import ProductUrlForm from "./ProductUrlForm";
-
-export default ProductUrlForm;
